Fix whitespace bug when expanding contractions
On the website demos:
https://github.com/nlp-compromise/website/blob/gh-pages/demos/demos.jsx#L211

The sentence:

We've only just begun, to live. White lace and promises. We'll start out
walkin' and learn to run.

is incorrectly expanded to:

We haveonly just begun, to live. White lace and promises. We willstart
out walkin' and learn to run.

(notice the whitespace errors). This commit adds trailing whitespace to
the expanded verbs resulting in:

We have only just begun, to live. White lace and promises. We will start
out walkin' and learn to run.

Added regression test.