Fix file/directory name case
The code references don't always match the actual filenames in case.  For example, `./src/index.js` refers to `./World/World`, but the actual path is `./world/World(.js)`.  This may work on macOS and Windows (which are case-insensitive by default), but not on Linux or on other case-sensitive systems.

This PR renames the files/directories based on what appears to be their canonical names (based on the source references and other directories).  Only two files (`scripts/pack.js` and `src/Doc/methods/output/_debug.js`) needed to be updated/edited to work with the new names.

Running `npm test` now succeeds 100% with no error messages on my system (WSL/ext4 on Windows), once I'd `npm install`'d in all of the plugin directories.

Fixes #649