Selectively ignore ".start" and ".end" matching token properties when handling a greedy match.
When handling a greedy match that's anchored to the start (`^`) or end (`$`), it is acceptable (normal, in fact!) for some of the matched terms to **not** be the first or last term in the phrase.  This fix attempts to allow that without accidentally forgetting that the greedy match as a whole should still either start or end the document.

Fixing this caused the `Doc.verbs()` behavior at https://github.com/spencermountain/compromise/blob/master/src/Subset/verbs/index.js#L15-L16 to change slightly (and correctly, I'd argue), such that a small change was needed at https://github.com/spencermountain/compromise/blob/master/src/Subset/verbs/methods.js#L45 (and on line 50) to compensate.

Added a test to specifically check for the anchored-greedy-match case.

(Note that this was fixed assuming PR #650, so that I could run the tests.)

Fixes #648