First-pass GitHub Actions workflow for building/testing compromise
This is an attempt at a very basic build/test workflow for GitHub Actions that can be triggered on any push (to master, or any branches or PRs).  The set of tests run may be excessive, as it wasn't clear to me what the difference between the various tests listed in `scripts` in `pacakge.json` meant.

In addtion, some missing devDependencies (eslint, nyc, tap-spec, ts-node, and typescript) have been added to ensure that the tests can all run (even if they aren't all passing at the moment).

Finally, a new `scripts/plugins.js` was added to perform arbitrary commands for each plugin.  The hope is that `scripts/plugin-check.js` can eventually be deprecated and replaced by (for example) `npm run plugins:build`.