Add "release" workflow
This is a very simple workflow to attempt to automate publishing releases to npm.  I was hoping to drive the version number in `package.json` from the GitHub release, but `npm version` would attempt to create a git tag which would conflict with the one created directly by the GitHub release process.

**Requirement:** You must have updated the intended version number in `package.json` _before_ you create the matching release in GitHub.  This will ensure that `package.json` tagged in release `X.Y.Z` says `"version": "X.Y.Z"`, and that version will also be used when publishing to npm.

**TO DO:** Three additional steps need to happen to make this work:
1. You'll need to create an npm authentication token via https://www.npmjs.com/settings/spencermountain/tokens.
2. You'll need to add this token as a secret named `NPM_TOKEN` at https://github.com/spencermountain/compromise/settings/secrets.
3. If you currently require 2FA to publish (at https://www.npmjs.com/package/compromise/access), you'll need to turn it off.

(The 'build-and-test' workflow was tweaked to make it easier to find failing steps... instead of having to sift through ten separate npm commands, they are bucketed into rough groups: install, lint, build, and test; each one of which is a small number of underlying npm commands.)