prevent doc offset changes (length) from clobbering term[0] offset
The original code assigned to the doc-level '.offset' result from
the offset information from term[0].  This creates an additional
reference to the same undelying object, *not* a clone of the data.
When the doc-level length was then set, it clobbered the length of
the term[0] offset info.

This now uses `Object.assign()` to create an all-new doc-level
offset object with the same values as it had previously.