Fix doc/result offset start
In cases where leading punctuation is included in a result/doc, we need to account for that difference from the `terms[0].offset.start` position.  The easiest way to do this is simply to subtract any leading positions by looking for how far into the result's `text` that `terms[0].text` starts.

Fixes #774