attempt to provide plugin-level options for compromise-dates
In particular, I have a use-case that needs `numbers().normalize()` _**not**_ to
run globally just because we're attempting to tag dates in the input.

The only way I could come up with to provide a hook to set any kind of plugin-
level configuration was to graft it on as a property of the main export.  If the
config hook is not used, the old behavior should be identical to what it was
before.  If compromise is extended as:

```javascript
const nlp = require('compromise');
const numbers = require('compromise-numbers');
const dates = require('compromise-dates');

nlp
  .extend(numbers)
  .extend(dates.config({ normalizeBeforeTagging: false }));
//             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^----- this is new!
```

... then compromise-dates will **not** normalize (numbers) before tagging.