Improve typescript cjs support for dates
The `compromise-dates` plugin was building a commonjs package, but the
types weren't properly exported in a way that TypeScript was able to
import them.

By providing an `index.d.cts` file and updating the `exports` property
of the `package.json` for this plugin we're able to now properly import
`compromise-dates` in a commonjs TypeScript project.